+++ scriptconfig/__init__.py
@@ -141,7 +141,7 @@ Ignore:
     mkinit ~/code/scriptconfig/scriptconfig/__init__.py --nomods --relative -w
 """
 
-__version__ = '0.7.5'
+__version__ = '0.7.6'
 
 __submodules__ = {
     'config': None,